pipeline/requirements.txt
# Core Dependencies
google-generativeai>=0.3.0
pymupdf>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0

# Data Handling
pandas>=2.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyarrow>=15.0.0
pyyaml>=6.0.0

# Utilities
tenacity>=8.2.0  # Retry logic for API calls
tqdm>=4.66.0  # Progress bars

# Development
pytest>=8.0.0
black>=24.0.0
ruff>=0.1.0
